fix register
UI done without function
